purplelinkweb/requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
pycryptodome==3.23.0
Werkzeug==3.1.3
flask
pycryptodome
bcrypt
mysql-connector-python
Flask-SocketIO==5.3.6
python-socketio==5.8.0
simple-websocket==0.10.0
dotenv==1.0.1